[IMP] partnerfirstname: introduce firstname.mixin Mutualize common things used by models that have firstname / lastname fields.
