+++ partner_firstname/models/__init__.py
@@ -1,3 +1,4 @@
 from . import base_config_settings
+from . import firstname_mixin
 from . import res_partner
 from . import res_users
