+++ partner_firstname/static/description/index.html
@@ -432,6 +432,8 @@ contact).</p>
 Maybe this configuration could depend on partner language, country or
 company, as discussed at <a class="reference external" href="https://github.com/OCA/partner-contact/issues/210">this OCA
 issue</a></p>
+<p>Move <tt class="docutils literal">copy()</tt> functions defined at <tt class="docutils literal">res.partner</tt> and <tt class="docutils literal">res.users</tt>
+level in firstname.mixin level, harmonizing different implementations.</p>
 </div>
 <div class="section" id="bug-tracker">
 <h2><a class="toc-backref" href="#toc-entry-4">Bug Tracker</a></h2>
